devtools/server/tests/mochitest/test_animation_actor-lifetime.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1247243
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1247243</title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();

  let gWalker = null;
  let animationsFront = null;

  addTest(async function setup() {
    info("Setting up inspector and animation actors.");

    const url = document.getElementById("animationContent").href;

    const { target } = await attachURL(url);
    const inspector = target.getInspector();
    animationsFront = target.getFront("animations");
    gWalker = await inspector.getWalker();
    ok(gWalker, "getWalker() should return an actor.");
    runNextTest();
  });

  addAsyncTest(async function testActorLifetime() {
    info("Testing animated node actor");
    const animatedNodeActor = await gWalker.querySelector(gWalker.rootNode,
      ".animated");
    await animationsFront.getAnimationPlayersForNode(animatedNodeActor);

    const animationsActor = DebuggerServer
                          .searchAllConnectionsForActor(animationsFront.actorID);

    is(animationsActor.actors.length, 1,
      "AnimationActor have 1 AnimationPlayerActors");

    info("Testing AnimationPlayerActors release");
    const stillNodeActor = await gWalker.querySelector(gWalker.rootNode,
      ".still");
    await animationsFront.getAnimationPlayersForNode(stillNodeActor);
    is(animationsActor.actors.length, 0,
      "AnimationActor does not have any AnimationPlayerActors anymore");

    info("Testing multi animated node actor");
    const multiNodeActor = await gWalker.querySelector(gWalker.rootNode,
      ".multi");
    await animationsFront.getAnimationPlayersForNode(multiNodeActor);
    is(animationsActor.actors.length, 2,
      "AnimationActor has now 2 AnimationPlayerActors");

    info("Testing single animated node actor");
    await animationsFront.getAnimationPlayersForNode(animatedNodeActor);
    is(animationsActor.actors.length, 1,
      "AnimationActor has only one AnimationPlayerActors");

    info("Testing AnimationPlayerActors release again");
    await animationsFront.getAnimationPlayersForNode(stillNodeActor);
    is(animationsActor.actors.length, 0,
      "AnimationActor does not have any AnimationPlayerActors anymore");

    runNextTest();
  });

  runNextTest();
};
  </script>
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1247243">Mozilla Bug 1247243</a>
  <a id="animationContent" target="_blank" href="animation-data.html">Test Document</a>
</body>
</html>
